bug fix in fglm (degenerate situation)

+++ src/msolve/msolve.c
@@ -550,7 +550,7 @@ static int add_random_linear_form_to_input_system(
     if (gens->field_char > 0) {
       int j = 0;
       for (i = len_old; i < len_new; ++i) {
-        gens->random_linear_form[j] = ((int32_t)(rand()/* pow(k, bcf) */) % gens->field_char);
+        gens->random_linear_form[j] = ((int16_t)(rand()) % gens->field_char);
         gens->cfs[i]  = gens->random_linear_form[j];
         k++;
         j++;
@@ -558,8 +558,9 @@ static int add_random_linear_form_to_input_system(
     }
     else {
       int j = 0;
+
       for (i = 2*len_old; i < 2*len_new; i += 2) {
-        gens->random_linear_form[j] = ((int32_t)(rand()));
+        gens->random_linear_form[j] = ((int16_t)(rand()));
         mpz_set_ui(*(gens->mpz_cfs[i]), gens->random_linear_form[j]);
         k++;
         j++;
@@ -738,7 +739,7 @@ check_and_set_vars_squared_in_monomial_basis(uint64_t *squvars,
     la i-ieme entree de squvars est a 0 si le monome vars[i]^2 n'est
     pas dans la base monomiale.
     Sinon, on met l'indice du monome. De toute facon, en 0, on aurait
-    le monome 1. 
+    le monome 1.
   */
   for (long i = 0; i < dquot; i++) {
     long deg = 0;
@@ -2874,6 +2875,7 @@ int msolve_trace_qq(mpz_param_t mpz_param,
   while(is_lucky_prime_ui(prime, bs_qq)){
     prime = next_prime(rand() % (1303905301 - (1<<30) + 1) + (1<<30));
   }
+
   uint32_t primeinit = prime;
   lp->p[0] = primeinit;
